+++ components/avatar/index.js
@@ -1,8 +1,20 @@
-import Styled from 'styled-components';
+import React from 'react';
+import { Box } from 'blockstack-ui';
 
-export default Styled.img`
-  border-radius: 50%;
-  max-width: ${({ size }) => size};
-  max-height: ${({ size }) => size};
-  max-width: 100%;
-`;
+
+const Avatar = ({ username, ...rest }) => (
+  <Box
+    size="42px"
+    display="block"
+    width="100%"
+    background={`#f8a5c2 url(${`/api/avatar/${username}`}) center center no-repeat`}
+    borderRadius="100%"
+    overflow="hidden"
+    style={{
+      backgroundSize: 'cover',
+    }}
+    {...rest}
+  />
+);
+
+export {Avatar}
